+++ components/panda/solid/src/stories/tree-view.stories.tsx
@@ -1,4 +1,4 @@
-import type { Meta } from '@storybook/react'
+import type { Meta } from 'storybook-solidjs'
 import { TreeView, type TreeViewData } from '~/components/ui'
 
 const meta: Meta = {